backend/requirements.txt
Django==2.2.28
djangorestframework==3.11.2
filetype
requests
python-dateutil
pillow==9.0.1
django-cors-headers
djangorestframework-filters